requirements.txt
fastapi
pyyaml
requests
wand